Backend/online_bookshaop/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
class-registry==2.1.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
distlib==0.3.9
Django==5.2.6
django-cors-headers==4.7.0
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
Faker==37.6.0
filelock==3.18.0
filters==1.3.2
filters-django==1.0.5
inflection==0.5.1
iniconfig==2.1.0
kombu==5.5.4
packaging==25.0
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2==2.9.10
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
regex==2024.11.6
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
virtualenv==20.34.0
wcwidth==0.2.13
